AeonFosil/AeonProject2/requirements.txt
customtkinter
pygame
edge-tts
ollama
groq
requests
beautifulsoup4
feedparser
googlesearch-python
pygetwindow
pyautogui
pillow
psutil
pyttsx3